Modified the GeotiffInputFormat to use a generic BinaryFileInputFormat
